refactor: extract UI components and simplify styling
