+++ src/jsMain/kotlin/ScanResult.kt
@@ -1,0 +1,1 @@
+data class ScanResult(val text: String, val format: Int)
